backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.5.2
fastapi-mcp==0.4.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
anthropic>=0.40.0
pandas==2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
requests>=2.28.0
rapidfuzz==3.6.1
python-dateutil==2.8.2
pypdf==4.0.2
reportlab>=4.0.0
pillow>=10.0.0
fpdf2==2.7.8
openpyxl==3.1.2
xlrd==2.0.1
supabase==2.3.0
postgrest==0.13.2
